fix: layout shift in loading indicator
